# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for Kuro API",
        "title": "Shop E-commerce ERP",
        "contact": {
            "name": "API Support",
            "email": "hoaiphong01012002@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/categories/get/name": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get by id product_category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Get by name product_category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/inventory/check/{warehouse_id}/{product_id}/{required_quantity}": {
            "get": {
                "description": "Check if the specified quantity of a product is available in a given warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Check inventory availability",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Warehouse ID",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required quantity",
                        "name": "required_quantity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/inventory/{_id}": {
            "get": {
                "description": "Retrieve an inventory record using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Get inventory by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inventory ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "Update an inventory item using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Update inventory by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inventory ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Inventory Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inventory_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/purchase_orders/create": {
            "post": {
                "description": "Create a new purchase order with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PurchaseOrder"
                ],
                "summary": "Create a new purchase order",
                "parameters": [
                    {
                        "description": "Purchase Order Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/purchase_order_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/purchase_orders/get/all": {
            "get": {
                "description": "Retrieve a list of purchase orders with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PurchaseOrder"
                ],
                "summary": "Get all purchase orders with pagination",
                "parameters": [
                    {
                        "type": "string",
                        "default": "1",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/purchase_orders/update/{_id}": {
            "put": {
                "description": "Update the details of a purchase order using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PurchaseOrder"
                ],
                "summary": "Update a purchase order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase Order ID",
                        "name": "_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Purchase Order Input",
                        "name": "purchase_order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/purchase_order_domain.Input"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "inventory_domain.Input": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
        "purchase_order_domain.Input": {
            "type": "object",
            "properties": {
                "order_date": {
                    "type": "string"
                },
                "order_number": {
                    "type": "string"
                },
                "status": {
                    "description": "Example: \"Processing\", \"Received\", \"Cancelled\"",
                    "type": "string"
                },
                "supplier": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
    },
...
}